Move esp-loader to seperatly compiled library
