+++ lib/esp-loader/README.md
@@ -1,0 +1,1 @@
+This source is from https://github.com/espressif/esp-serial-flasher
