+++ lib/esp-loader/private_include/esp_targets.h
